requirements.txt
# 可能有一些遗漏

torch
pyyaml
pandas
numpy
tqdm
matplotlib